Fixed some bugs with the individual battle logic implementation

+++ src/individual/state.jl
@@ -22,18 +22,18 @@ struct DynamicIndividualState
 end
 
 has_shields(state::DynamicIndividualState, agent::Int64) = agent == 1 ?
-    iszero(state.data % 3) : iszero((state.data ÷ 3) % 3)
+  iszero(state.data % UInt32(3)) : iszero((state.data ÷ UInt32(3)) % UInt32(3))
 
 get_fast_moves_pending(state::DynamicIndividualState) =
-    (state.data ÷ 9) % 7, (state.data ÷ 63) % 7
+    (state.data ÷ UInt32(9)) % UInt32(7), (state.data ÷ UInt32(63)) % UInt32(7)
 
-get_cmp(state::DynamicIndividualState) = (state.data ÷ 441) % 5
+get_cmp(state::DynamicIndividualState) = (state.data ÷ UInt32(441)) % UInt32(5)
 
-get_chance(state::DynamicIndividualState) = (state.data ÷ 2205) % 6
+get_chance(state::DynamicIndividualState) = (state.data ÷ UInt32(2205)) % UInt32(6)
 
 get_buffs(state::DynamicIndividualState, attacker::Int64) = attacker == 1 ?
-    ((state.data ÷ 13230)   % 9, (state.data ÷ 119070)  % 9) :
-    ((state.data ÷ 1071630) % 9, (state.data ÷ 9644670) % 9)
+    ((state.data ÷ UInt32(13230))   % 9, (state.data ÷ UInt32(119070))  % 9) :
+    ((state.data ÷ UInt32(1071630)) % 9, (state.data ÷ UInt32(9644670)) % 9)
 
 DynamicIndividualState(state::StaticIndividualState; shields::Int8 = Int8(2)) = DynamicIndividualState(
     DynamicPokemon.(state.teams),
